Move LightboxFooter into the implementation file
